virt console: extract the popup code to a separate component
